requirements.txt
google-genai>=1.5.0
opencv-python>=4.8.0
pyaudio>=0.2.14
pillow>=10.0.0
mss>=10.0.0
SpeechRecognition>=3.10.0
pydub>=0.25.1
numpy>=1.24.0
python-dotenv>=1.0.0